Exclude libjingle_peerconnection_{jni,so} targets from Chromium builds.
In GN, the libjingle_peerconnection_jni target becomes a part of
'all' implicitly, which surfaced the incompability between it
and the Chromium logging implementation. In the GYP build, the
target is not present due to api.gyp not being depended upon yet.

BUG=webrtc:4256
TBR=perkj@webrtc.org
NOTRY=True

Review-Url: https://codereview.webrtc.org/2082573004
Cr-Original-Commit-Position: refs/heads/master@{#13231}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 69b34625c1ee2f527e481c420ff1294851463b77