SurfaceViewRendererOnMeasureTest: Wait for frame size change to take effect
SurfaceViewRendererOnMeasureTest#testFrame1280x720 is currently flaky
because of a race. This issue was introduced here:
https://codereview.webrtc.org/2111933002/. That CL moved the function
updateFrameDimensionsAndReportEvents() from renderFrame() to
renderFrameOnRenderThread(). The OnMeasureTest currently just calls
renderFrame() with a new frame size and immediately starts testing the
measured size, which might be before renderFrameOnRenderThread() is
executed. This CL waits for the RendererEvents.onFrameResolutionChanged()
callback before starting the test to fix this race.

BUG=webrtc:6089

Review-Url: https://codereview.webrtc.org/2147463002
Cr-Original-Commit-Position: refs/heads/master@{#13448}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: e5a246f69341007a69f5244acac86e218e341bfe