Don't use LOG_END timestamp to compute RtcEventLog duration.
The LOG_END time is incorrect if the event log is stopped
by the file size limit instead of a duration limit or a
manual stop. This makes the call appear to be very long
(10^13 seconds) in the analysis tool. This CL is a workaround
for that problem.

BUG=webrtc:6138

Review-Url: https://codereview.webrtc.org/2176663002
Cr-Original-Commit-Position: refs/heads/master@{#13585}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 88c1d2bc0f8468f04cb05a536c5194f9bb9d0e50